Modify: separate test and sentinel
