--- support/locust/credentials.py
@@ -1,6 +1,0 @@
-def gen_username(id: int) -> str:
-    return f"jaclang{id}@jaseci.org"
-
-
-def gen_password(id: int) -> str:
-    return f"ilovejaclang{id}"
